+++ mutt_logging.c
@@ -271,7 +271,7 @@ int level_validator(const struct ConfigSet *cs, const struct ConfigDef *cdef,
 {
   if ((value < 0) || (value >= LL_MAX))
   {
-    buf_printf(err, _("Invalid value for option %s: %ld"), cdef->name, value);
+    buf_printf(err, _("Invalid value for option %s: %ld"), cdef->name, (long) value);
     return CSR_ERR_INVALID;
   }
